Hack: Add db:migrate and db:seed capistrano tasks
Required for initial db setup for poc. Capistrano-rails tasks won't just
work out of the box as rails app is in a sub-directory of deployment
directory, and therefore Gemfile and .bundle directory won't be found.

I'm sure it could be made to work, but this is good enough for a poc.